chore: modularize code
