+++ frontend/src/routes/checkout-failed.tsx
@@ -11,24 +11,12 @@ import {
 import { CloseIcon } from "@chakra-ui/icons"
 import { AnimatedCard } from "../components/Common/AnimatedCard"
 import { CircleIcon } from "../components/Common/CircleIcon"
-import { useEffect, useState } from "react"
-import { CheckoutService } from "../client/sdk.gen"
 import useAuth from "../hooks/useAuth"
-import useCustomToast from "../hooks/useCustomToast"
+import { useProcessFailedCheckout } from "../hooks/useProcessFailedCheckout"
 
 // Define search params
 interface CheckoutFailedSearchParams {
   sessionId?: string
-}
-
-// Define error interface for proper typing
-interface ApiError {
-  response?: {
-    status: number
-  }
-  status?: number
-  name?: string
-  message?: string
 }
 
 export const Route = createFileRoute("/checkout-failed")({
@@ -43,61 +31,20 @@ export const Route = createFileRoute("/checkout-failed")({
 function CheckoutFailed() {
   const { sessionId } = Route.useSearch()
   const { user: authUser, isLoading } = useAuth()
-  const showToast = useCustomToast()
   const navigate = useNavigate()
-  const [shouldShowCard, setShouldShowCard] = useState(false)
+  
+  // Use our custom hook for failed checkout processing
+  const { shouldShowCard } = useProcessFailedCheckout({
+    authUser,
+    isLoading,
+    sessionId
+  })
   
   // Colors for theming
   const cardBgColor = useColorModeValue("white", "#2D3748")
   const buttonBgColor = useColorModeValue("black", "#00766c")
   const errorBgColor = useColorModeValue("red.500", "red.600")
   const textColor = useColorModeValue("gray.800", "white")
-
-  useEffect(() => {
-    if (!isLoading && !authUser) {
-      navigate({ to: "/login" })
-      return
-    }
-
-    if (!sessionId) {
-      navigate({ to: "/" })
-      return
-    }
-
-    const processFailedCheckout = async () => {
-      try {
-        // Process the failed checkout with stripeCancel
-        await CheckoutService.stripeCancel({ sessionId })
-        // If the stripeCancel call succeeds, show the failure card
-        setShouldShowCard(true)
-      } catch (error: unknown) {
-        console.error("Error processing failed checkout:", error)
-        
-        // More robust error checking for 404 responses
-        const apiError = error as ApiError
-        
-        // Check different possible ways a 404 might be represented
-        const is404 = 
-          (apiError.response && apiError.response.status === 404) || 
-          apiError.status === 404 || 
-          (apiError.message && apiError.message.includes("404")) ||
-          (apiError.name && apiError.name.includes("NotFound"));
-          
-        if (is404) {
-          console.log("404 error detected, redirecting to dashboard");
-          // If 404, redirect directly to dashboard without showing any modal
-          navigate({ to: "/" })
-        } else {
-          // For other errors, still show the failure screen to give feedback
-          setShouldShowCard(true)
-        }
-      }
-    }
-
-    if (authUser && sessionId) {
-      processFailedCheckout()
-    }
-  }, [authUser, sessionId, isLoading, navigate, showToast])
 
   if (!shouldShowCard) {
     return (
